+++ src/lib/combinators.rs
@@ -60,8 +60,8 @@ impl From<bool> for FlowControl {
 /// assert_eq!(parse(parse_xes, input3).result, None);
 /// ```
 #[inline]
-pub fn bind<I:SliceLike, O1, O2, P, S>(p: impl Parser<I, O1, S>,
-                                       f: impl FnOnce(O1) -> P + Copy
+pub const fn bind<I:SliceLike, O1, O2, P, S>(p: impl Parser<I, O1, S>,
+                                             f: impl FnOnce(O1) -> P + Copy
 ) -> impl Parser<I, O2, S> where P: Parser<I, O2, S> {
     create_parser!(s, f(p(s)?)(s))
 }
@@ -94,7 +94,7 @@ pub fn bind<I:SliceLike, O1, O2, P, S>(p: impl Parser<I, O1, S>,
 /// assert_eq!(parse(parse_next_even, input3).result, Some(4));
 /// ```
 #[inline]
-pub fn map<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
+pub const fn map<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
                                    f: impl FnOnce(O) -> O2 + Copy
 ) -> impl Parser<I, O2, S> {
     map!(f, p)
@@ -133,7 +133,7 @@ pub fn map<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
 /// assert_eq!(parse(parse_binary, input3).result, None);
 /// ```
 #[inline]
-pub fn map_if<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
+pub const fn map_if<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
                                       f: impl FnOnce(O) -> Option<O2> + Copy
 ) -> impl Parser<I, O2, S> {
     create_parser!(s, { p(s).and_then(f) })
@@ -163,7 +163,7 @@ pub fn map_if<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
 /// assert_eq!(result1, Some("1234".to_owned()));
 /// ```
 #[inline]
-pub fn into_type<I: SliceLike, O: Into<T>, T, S>(p: impl Parser<I, O, S>) -> impl Parser<I, T, S> {
+pub const fn into_type<I: SliceLike, O: Into<T>, T, S>(p: impl Parser<I, O, S>) -> impl Parser<I, T, S> {
     map!(O::into, p)
 }
 
@@ -190,8 +190,8 @@ pub fn into_type<I: SliceLike, O: Into<T>, T, S>(p: impl Parser<I, O, S>) -> imp
 /// assert_eq!(parse(parse_even, input2).result, Some(2));
 /// ```
 #[inline]
-pub fn filter<I: SliceLike, O, S>(p: impl Parser<I, O, S>,
-                                  f: impl FnOnce(&O) -> bool + Copy
+pub const fn filter<I: SliceLike, O, S>(p: impl Parser<I, O, S>,
+                                        f: impl FnOnce(&O) -> bool + Copy
 ) -> impl Parser<I, O, S> {
     create_parser!(s, p(s).filter(f))
 }
@@ -218,7 +218,7 @@ pub fn filter<I: SliceLike, O, S>(p: impl Parser<I, O, S>,
 /// assert_eq!(parse(parse_optional_int, input2).result, Some(None));
 /// ```
 #[inline]
-pub fn succeed<I:SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, Option<O>, S> {
+pub const fn succeed<I:SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, Option<O>, S> {
     create_parser!(s, {
         Some(p(s))
     })
@@ -245,7 +245,7 @@ pub fn succeed<I:SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, Opt
 /// assert_eq!(result.state, input);
 /// ```
 #[inline]
-pub fn peek<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
+pub const fn peek<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
     create_parser!(s, {
         let pos = s.input;
         let res = p(s);
@@ -274,7 +274,7 @@ pub fn peek<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S>
 /// assert_eq!(parse(parse_digits, input2).result, None);
 /// ```
 #[inline]
-pub fn not_empty<I: SliceLike, O: SliceLike, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
+pub const fn not_empty<I: SliceLike, O: SliceLike, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
     filter(p, |r| !r.slice_is_empty())
 }
 
@@ -304,7 +304,7 @@ pub fn not_empty<I: SliceLike, O: SliceLike, S>(p: impl Parser<I, O, S>) -> impl
 /// assert_eq!(result2.state, "4321");
 /// ```
 #[inline]
-pub fn attempt<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
+pub const fn attempt<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O, S> {
     create_parser!(s, {
         let pos = s.input;
         let res = p(s);
@@ -336,7 +336,7 @@ pub fn attempt<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, O,
 /// assert_eq!(parse(parse_int, input).result, Some((4, 1234)));
 /// ```
 #[inline]
-pub fn count_consumed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, (I::Idx, O), S> {
+pub const fn count_consumed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, (I::Idx, O), S> {
     create_parser!(s, {
         let old = s.input.slice_len();
         let res = p(s)?;
@@ -364,7 +364,7 @@ pub fn count_consumed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parse
 /// assert_eq!(parse(parse_int, input).result, Some((input, 1234)));
 /// ```
 #[inline]
-pub fn and_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, (I, O), S> {
+pub const fn and_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, (I, O), S> {
     create_parser!(s, {
         let old_input = s.input;
         let res = p(s)?;
@@ -392,7 +392,7 @@ pub fn and_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I,
 /// assert_eq!(parse(parse_abc_then_123, input).result, Some(input));
 /// ```
 #[inline]
-pub fn get_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, I, S> {
+pub const fn get_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I, I, S> {
     create_parser!(s, {
         let old_input = s.input;
         p(s)?;
@@ -422,7 +422,7 @@ pub fn get_parsed<I: SliceLike, O, S>(p: impl Parser<I, O, S>) -> impl Parser<I,
 /// assert_eq!(parse(parse_4_digits, input2).result, None);
 /// ```
 #[inline]
-pub fn times<I: SliceLike, O, S>(times: u32, p: impl Parser<I, O, S>) -> impl Parser<I, I, S> {
+pub const fn times<I: SliceLike, O, S>(times: u32, p: impl Parser<I, O, S>) -> impl Parser<I, I, S> {
     create_parser!(s, {
         let old_input = s.input;
 
@@ -454,8 +454,8 @@ pub fn times<I: SliceLike, O, S>(times: u32, p: impl Parser<I, O, S>) -> impl Pa
 /// assert_eq!(parse(parse_abc_then_123, input).result, Some("123"));
 /// ```
 #[inline]
-pub fn right<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
-                                      p2: impl Parser<I, O2, S>
+pub const fn right<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
+                                            p2: impl Parser<I, O2, S>
 ) ->  impl Parser<I, O2, S> {
     create_parser!(s, {
         p1(s).and_then(|_| p2(s))
@@ -484,8 +484,8 @@ pub fn right<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
 /// assert_eq!(parse(parse_abc_then_123, input).result, Some("abc"));
 /// ```
 #[inline]
-pub fn left<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
-                                     p2: impl Parser<I, O2, S>
+pub const fn left<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
+                                           p2: impl Parser<I, O2, S>
 ) ->  impl Parser<I, O1, S> {
     create_parser!(s, {
         p1(s).and_then(|res| p2(s).map(|_| res))
@@ -512,9 +512,9 @@ pub fn left<I: SliceLike, S, O1, O2>(p1: impl Parser<I, O1, S>,
 /// assert_eq!(parse(parse_middle, input).result, Some("123"));
 /// ```
 #[inline]
-pub fn middle<I: SliceLike, S, O1, O2, O3>(p1: impl Parser<I, O1, S>,
-                                           p2: impl Parser<I, O2, S>,
-                                           p3: impl Parser<I, O3, S>
+pub const fn middle<I: SliceLike, S, O1, O2, O3>(p1: impl Parser<I, O1, S>,
+                                                 p2: impl Parser<I, O2, S>,
+                                                 p3: impl Parser<I, O3, S>
 ) ->  impl Parser<I, O2, S> {
     right(p1, left(p2, p3))
 }
@@ -551,8 +551,8 @@ macro_rules! internal_or {
         /// assert_eq!(parse(parse_abc_or_123, input3).result, None);
         /// ```
         #[inline]
-        pub fn $id<I: SliceLike, O, S>(p1: impl Parser<I, O, S>,
-                                       p2: impl Parser<I, O, S>
+        pub const fn $id<I: SliceLike, O, S>(p1: impl Parser<I, O, S>,
+                                             p2: impl Parser<I, O, S>
         ) -> impl Parser<I, O, S> {
             create_parser!(s, {
                 let pos = s.input;
@@ -605,8 +605,8 @@ macro_rules! internal_or_diff {
         /// assert_eq!(parse(parse_int_or_abc, input3).result, None);
         /// ```
         #[inline]
-        pub fn $id<I: SliceLike, O1, O2, S>(p1: impl Parser<I, O1, S>,
-                                            p2: impl Parser<I, O2, S>
+        pub const fn $id<I: SliceLike, O1, O2, S>(p1: impl Parser<I, O1, S>,
+                                                  p2: impl Parser<I, O2, S>
         ) -> impl Parser<I, (), S> {
             create_parser!(s, {
                 let pos = s.input;
@@ -651,8 +651,8 @@ internal_or_diff!(or_diff_no_partial, false, "This differs from `or_diff` in tha
 /// assert_eq!(nums, vec![123]);
 /// ```
 #[inline]
-pub fn lift_to_state<I: SliceLike, S, O1, O2>(f: impl FnOnce(&mut S, O1) -> O2 + Copy,
-                                              p: impl Parser<I, O1, S>
+pub const fn lift_to_state<I: SliceLike, S, O1, O2>(f: impl FnOnce(&mut S, O1) -> O2 + Copy,
+                                                    p: impl Parser<I, O1, S>
 ) -> impl Parser<I, O2, S> {
     create_parser!(s, {
         p(s).map(|res| f(s.user_state, res))
@@ -666,7 +666,7 @@ pub fn lift_to_state<I: SliceLike, S, O1, O2>(f: impl FnOnce(&mut S, O1) -> O2 +
 /// * `p` - a parser for the separator
 /// * `allow_trailing` - whether a trailing separator is allowed.
 #[inline]
-pub fn separator<I, O, S>(p: impl Parser<I, O, S>, allow_trailing: bool) -> Option<(bool, impl Parser<I, O, S>)> {
+pub const fn separator<I, O, S>(p: impl Parser<I, O, S>, allow_trailing: bool) -> Option<(bool, impl Parser<I, O, S>)> {
     Some((allow_trailing, p))
 }
 
@@ -674,7 +674,7 @@ pub fn separator<I, O, S>(p: impl Parser<I, O, S>, allow_trailing: bool) -> Opti
 /// argument when no separator should be present.
 #[allow(unreachable_code)]
 #[inline]
-pub fn no_separator<I: SliceLike, S>() -> Option<(bool, impl Parser<I, (), S>)> {
+pub const fn no_separator<I: SliceLike, S>() -> Option<(bool, impl Parser<I, (), S>)> {
     return None;
 
     // Unreachable, but provides type/size information about the return value
@@ -710,7 +710,7 @@ fn many_internal<I: SliceLike, O, O2, S, F: Into<FlowControl>>(
         }
     }
 
-    separator.is_none_or(|(allow_trailing, _)| allow_trailing || !has_trailing_sep)
+    !separator.is_some_and(|(allow_trailing, _)| !allow_trailing && has_trailing_sep)
         && (allow_empty || successes)
 }
 
@@ -739,9 +739,9 @@ fn many_internal<I: SliceLike, O, O2, S, F: Into<FlowControl>>(
 /// assert_eq!(parse(parse_nums, input).result, Some("1,2,3"));
 /// ```
 #[inline]
-pub fn many<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
-                                    allow_empty: bool,
-                                    separator: Option<(bool, impl Parser<I, O2, S>)>,
+pub const fn many<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
+                                          allow_empty: bool,
+                                          separator: Option<(bool, impl Parser<I, O2, S>)>,
 ) -> impl Parser<I, I, S> {
     create_parser!(s, {
         let old_input = s.input;
@@ -782,7 +782,7 @@ pub fn many<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
 /// assert_eq!(parse(parse_nums, input).result, Some(6));
 /// ```
 #[inline]
-pub fn fold<I: SliceLike, O, O2, S, R, F: Into<FlowControl>>(
+pub const fn fold<I: SliceLike, O, O2, S, R, F: Into<FlowControl>>(
     p: impl Parser<I, O, S>,
     init: impl FnOnce() -> R + Copy,
     f: impl FnOnce(&mut R, O) -> F + Copy,
@@ -836,7 +836,7 @@ pub fn fold<I: SliceLike, O, O2, S, R, F: Into<FlowControl>>(
 /// assert_eq!(parse(parse_nums, input).result, None);
 /// ```
 #[inline]
-pub fn many_to_array<I: SliceLike, O, O2, S, const N: usize>(
+pub const fn many_to_array<I: SliceLike, O, O2, S, const N: usize>(
     p: impl Parser<I, O, S>,
     constructor: impl FnOnce() -> ([O; N], usize) + Copy,
     allow_empty: bool,
@@ -879,9 +879,10 @@ pub fn many_to_array<I: SliceLike, O, O2, S, const N: usize>(
 /// assert_eq!(parse(parse_nums, input).result, Some(vec![1,2,3]));
 /// ```
 #[inline]
-pub fn many_to_vec<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
-                                           allow_empty: bool,
-                                           separator: Option<(bool, impl Parser<I, O2, S>)>
+pub const fn many_to_vec<I: SliceLike, O, O2, S>(
+    p: impl Parser<I, O, S>,
+    allow_empty: bool,
+    separator: Option<(bool, impl Parser<I, O2, S>)>
 ) -> impl Parser<I, Vec<O>, S> {
     fold(p, Vec::new, |v, x| v.push(x), allow_empty, separator)
 }
@@ -921,9 +922,10 @@ pub fn many_to_vec<I: SliceLike, O, O2, S>(p: impl Parser<I, O, S>,
 /// assert_eq!(parse(parse_nums, input).result, Some(expected));
 /// ```
 #[inline]
-pub fn many_to_map<I: SliceLike, K: Hash + Eq, V, O2, S>(p: impl Parser<I, (K, V), S>,
-                                                         allow_empty: bool,
-                                                         separator: Option<(bool, impl Parser<I, O2, S>)>,
+pub const fn many_to_map<I: SliceLike, K: Hash + Eq, V, O2, S>(
+    p: impl Parser<I, (K, V), S>,
+    allow_empty: bool,
+    separator: Option<(bool, impl Parser<I, O2, S>)>,
 ) -> impl Parser<I, HashMap<K, V>, S> {
     fold(p, HashMap::new, |m, (k, v)| { m.insert(k, v); }, allow_empty, separator)
 }
@@ -942,9 +944,10 @@ pub fn many_to_map<I: SliceLike, K: Hash + Eq, V, O2, S>(p: impl Parser<I, (K, V
 /// ### Example
 /// See [`many_to_map`]
 #[inline]
-pub fn many_to_map_ordered<I: SliceLike, K: Ord, V, O2, S>(p: impl Parser<I, (K, V), S>,
-                                                           allow_empty: bool,
-                                                           separator: Option<(bool, impl Parser<I, O2, S>)>,
+pub const fn many_to_map_ordered<I: SliceLike, K: Ord, V, O2, S>(
+    p: impl Parser<I, (K, V), S>,
+    allow_empty: bool,
+    separator: Option<(bool, impl Parser<I, O2, S>)>,
 ) -> impl Parser<I, BTreeMap<K, V>, S> {
     fold(p, BTreeMap::new, |m, (k, v)| { m.insert(k, v); }, allow_empty, separator)
 }
@@ -981,8 +984,8 @@ pub fn many_to_map_ordered<I: SliceLike, K: Ord, V, O2, S>(p: impl Parser<I, (K,
 /// assert_eq!(parse(parse_greedy, input2).result, Some(0));
 /// ```
 #[inline]
-pub fn greedy_or<I: SliceLike, S, O>(p1: impl Parser<I, O, S>,
-                                     p2: impl Parser<I, O, S>
+pub const fn greedy_or<I: SliceLike, S, O>(p1: impl Parser<I, O, S>,
+                                           p2: impl Parser<I, O, S>
 ) ->  impl Parser<I, O, S> {
     create_parser!(s, {
         let pos = s.input;
@@ -1074,8 +1077,8 @@ pub fn greedy_or<I: SliceLike, S, O>(p1: impl Parser<I, O, S>,
 /// assert_eq!(parse(expr(), input5).result, Some(-25));
 /// ```
 #[inline]
-pub fn chain<I: SliceLike, S, O, F>(p: impl Parser<I, O, S>,
-                                    op: impl Parser<I, F, S>
+pub const fn chain<I: SliceLike, S, O, F>(p: impl Parser<I, O, S>,
+                                          op: impl Parser<I, F, S>
 ) ->  impl Parser<I, O, S> where F: FnOnce(O, O) -> O {
     create_parser!(s, {
         let mut res = p(s)?;