+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add extension functions for starting a parse. Instead of writing
   e.g. `parse(p, input)` you can write `p.parse(input)`.
+- Made all parsers and combinators `const`. Note that parsing is _not_ `const` (yet).
 
 ## 0.9.0